tons of MSVC level4 warnings addressed; some useful; misc code clean ups and refactor

+++ libs/base/include/mrpt/math/CHistogram.h
@@ -83,7 +83,7 @@ namespace math
 		/** Retuns the elements count into the selected bin index, where first one is 0.
 		 * \exception std::exception On invalid index
 		 */
-		int		getBinCount(const size_t index) const;
+		size_t getBinCount(const size_t index) const;
 
 		/** Retuns the ratio in [0,1] range for the selected bin index, where first one is 0.
 		 *  It returns 0 if no elements have been added.
